# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LAB APP - API Documentation",
        "description": "API para gestão de escolas.",
        "contact": {
            "email": "pedro.manuel.dev@gmail.com"
        },
        "license": {
            "name": "License Name",
            "url": "http://url-to-license.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/receive-provinces": {
            "post": {
                "tags": [
                    "Provinces"
                ],
                "summary": "Faz a requisição e salva as provincias usando uma API externa.",
                "operationId": "fetchAndSaveProvinces",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "endpoint": {
                                        "description": "URL do endpoint que retorna os dados das provincias.",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Erro na requisição do endpoint",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Uma mensagem de erro",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Uma mensagem de sucesso",
                                            "type": "string",
                                            "example": "Provincias importadas com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}